+++ packages/core/ios/SentrySDKWrapper.h
@@ -1,6 +1,5 @@
 #import <Foundation/Foundation.h>
 
-@class SentryOptions;
 @class SentryScope;
 
 @interface SentrySDKWrapper : NSObject
@@ -13,22 +12,8 @@
 
 + (BOOL)crashedLastRun;
 
-+ (void)startWithOptions:(SentryOptions *)options;
-
-+ (SentryOptions *)createOptionsWithDictionary:(NSDictionary *)options
-                        isSessionReplayEnabled:(BOOL)isSessionReplayEnabled
-                                         error:(NSError **)errorPointer;
-
-+ (void)setupWithDictionary:(NSDictionary *)options
-     isSessionReplayEnabled:(BOOL)isSessionReplayEnabled
-                      error:(NSError **)errorPointer;
-
 + (BOOL)debug;
 
 + (NSString *)releaseName;
-
-+ (BOOL)enableAutoSessionTracking;
-
-+ (BOOL)enableWatchdogTerminationTracking;
 
 @end